+++ src/writeword.cc
@@ -41,7 +41,7 @@ NAN_METHOD(WriteWordAsync) {
   NanScope();
 
   if (args.Length() < 4 || !args[0]->IsInt32() || !args[1]->IsInt32() || !args[2]->IsInt32() || !args[3]->IsFunction()) {
-    return NanThrowError("incorrect arguments passed to writeWord(int fd, int cmd, int val, function cb)");
+    return NanThrowError("incorrect arguments passed to writeWord(int fd, int cmd, int word, function cb)");
   }
 
   int fd = args[0]->Int32Value();
@@ -57,7 +57,7 @@ NAN_METHOD(WriteWordSync) {
   NanScope();
 
   if (args.Length() < 3 || !args[0]->IsInt32() || !args[1]->IsInt32() || !args[2]->IsInt32()) {
-    return NanThrowError("incorrect arguments passed to writeWordSync(int fd, int cmd, int val)");
+    return NanThrowError("incorrect arguments passed to writeWordSync(int fd, int cmd, int word)");
   }
 
   int fd = args[0]->Int32Value();